Permit auth with developer token, remove sdk
Allow authentication using an evernote developer token
which can be obtained via
https://www.evernote.com/api/DeveloperToken.action
and set in configuration as
(setq evernote-developer-token "<devtoken>")

Changes follow the pattern from the patch at
https://github.com/kechako/emacs-evernote-mode-developer-token

As well, the included SDK in the ruby/lib directory
is removed, and local references are likewise removed
in enclient.rb since the SDK can be installed
using
gem install evernote-thrift

Note that the modifications to enclient.rb do not
properly handle the expiration of developer tokens,
since I can find neither a way to programmatically
determine their expiration date nor a way to
refresh to a new one.

A better implementation of all this a proper
implementation of OAuth, but that's beyond the scope
of this commit.